+++ CHANGELOG.md
@@ -1,5 +1,7 @@
+### v0.3.0 - 2023-??-??
+* Add `finite` validation for float types which checks against NaN and infinity.
 
-### v0.2.0 - 2023-??-??
+### v0.2.0 - 2023-04-13
 
 * [BREAKING] Rename string validator `present` -> `not_empty`. Rename error variant `Missing` -> `Empty`.
 * [BREAKING] Rename feature `serde1` to `serde`.